Fix PrintChoiceValue in the Ada backend
When importing a type from another ASN.1 module, and declaring a CHOICE
constant using a type from the other module, the generated code failed
to set the value of the constant, using the syntax (kind =>
xxxx_PRESENT, value=> ...). One issue was that the xxx_PRESENT is an
enumerated declared in a different Ada namespace (thus is not visible at
the place it is used) ; the other issue is that the type being declared
private, it is not possible to set it this way. The solution is to use
the setter that is public, and generated for each CHOICE alternative.