Generate ACN AST in HTML from template
The AST in HTML was hardcoded in Comments.cs, for both ASN.1 and ACN.
This commit is a refactoring - the hardcoded C# function was rewritten
in F# and uses the stringtemplate backend to generate the HTML code.
This enables the possibility of non-HTML backends for ICDs also
including the ACN models. (e.g. Latex)